Ignore user-defined AWS_PROFILE variable in cache uploads
When the S3 RoleARN config is used, someone could define a CI/CD
variable was AWS_PROFILE environment variable and cause the AWS SDK to
look up the profile from config files. If the profile does not exist,
the cache uploader fails with an error: `failed to get shared config
profile`.

To avoid this, make sure we pass an empty AWS_PROFILE variable to the
runner helper.

Relates to
https://gitlab.com/gitlab-com/request-for-help/-/work_items/3883