Ensure `stable docker images` only runs after all tests pass
Previously during a release the `stable docker images` job ran only
after the runner and helper images were built. However, this published
Docker images before the package was published to PackageCloud (`stable
packagecloud`) and S3 (`stable S3`).

If a test blocked the `test` stage from continuining, we are then left
in a state where Docker tags are available, but the binaries are not.

To avoid this issue, use `dependencies` instad of `needs` section so
that all release steps occur at the same time.

Relates to
https://gitlab.com/gitlab-com/runner-group/team-tasks/-/issues/431