Cleanup dangling virtualbox resources
When job is cancelled while the VM is being created,
the VM creation is aborted midway and it leaves dangling
resources which are not cleaned up. This prevents from a new job
getting scheduled in certain cases since the VM name matches.

To resolve this, we delete all the hard drives associated to the
VM and delete the machine folder where the VM data is present.